backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
cryptography==43.0.3
itsdangerous==2.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
